TST: Add tests for annotate.
